add pattern matching and rule application
There is a shortcoming to the pattern matching:
It does not handle repeated modifier variables

So for example "m1 Onion m1 Potato" will not be matched
properly because of the repeated m1.
This is not a big deal since I don't see us using rules with a repeated
variable on one side, but it would be nice to improve it.

To improve it I think we'd need a rewrite of util.match_pattern,
which you are welcome to try. It's a fun problem.
If you try it then remove the skip decorator from
test_pattern5 and make sure all tests pass.